requirements.txt
file-read-backwards==2.0.0
matplotlib==3.2.1
numpy==1.18.2
python-dateutil==2.8.1
scipy==1.4.1
